Dead Reckoning and Particle Filter now in Local Service.
Mapping still doesn't work. No longer suitable for Demo shows.

+++ app/src/main/java/com/example/krisma/architek/trackers/HeadingListener.java
@@ -5,6 +5,6 @@ package com.example.krisma.architek.trackers;
  */
 public interface HeadingListener {
 
-    void onHeadingChange(double heading);
+    void onHeadingChange(float heading);
 
 }
